# Improve WKB performance.